+++ npbench/benchmarks/go_fast/go_fast.py
@@ -6,5 +6,5 @@ import numpy as np
 def initialize(N):
     from numpy.random import default_rng
     rng = default_rng(42)
-    x = rng.random((N, N), dtype=np.float32)
+    x = rng.random((N, N), dtype=np.float64)
     return x
