+++ npbench/benchmarks/nbody/nbody_numba_op.py
@@ -94,8 +94,8 @@ def nbody(mass, pos, vel, N, Nt, dt, G, softening):
     acc = getAcc(pos, mass, G, softening)
 
     # calculate initial energy of system
-    KE = np.ndarray(Nt + 1, dtype=np.float32)
-    PE = np.ndarray(Nt + 1, dtype=np.float32)
+    KE = np.ndarray(Nt + 1, dtype=np.float64)
+    PE = np.ndarray(Nt + 1, dtype=np.float64)
     KE[0], PE[0] = getEnergy(pos, vel, mass, G)
 
     t = 0.0
