+++ npbench/benchmarks/polybench/adi/adi_dace.py
@@ -7,15 +7,15 @@ N = dc.symbol('N', dtype=dc.int64)
 
 
 @dc.program
-def kernel(TSTEPS: dc.int64, u: dc.float32[N, N]):
+def kernel(TSTEPS: dc.int64, u: dc.float64[N, N]):
 
     v = np.empty(u.shape, dtype=u.dtype)
     p = np.empty(u.shape, dtype=u.dtype)
     q = np.empty(u.shape, dtype=u.dtype)
 
-    DX = 1.0 / np.float32(N)
-    DY = 1.0 / np.float32(N)
-    DT = 1.0 / np.float32(TSTEPS)
+    DX = 1.0 / np.float64(N)
+    DY = 1.0 / np.float64(N)
+    DT = 1.0 / np.float64(TSTEPS)
     B1 = 2.0
     B2 = 1.0
     mul1 = B1 * DT / (DX * DX)
