+++ npbench/benchmarks/polybench/bicg/bicg_dace.py
@@ -5,6 +5,6 @@ M, N = (dc.symbol(s, dtype=dc.int64) for s in ('M', 'N'))
 
 
 @dc.program
-def kernel(A: dc.float32[N, M], p: dc.float32[M], r: dc.float32[N]):
+def kernel(A: dc.float64[N, M], p: dc.float64[M], r: dc.float64[N]):
 
     return r @ A, A @ p
