+++ npbench/benchmarks/polybench/cholesky2/cholesky2_dace.py
@@ -6,7 +6,7 @@ k = dc.symbol('k', dtype=dc.int64)
 
 
 @dc.program
-def triu(A: dc.float32[N, N]):
+def triu(A: dc.float64[N, N]):
     B = np.zeros_like(A)
     for i in dc.map[0:N]:
         for j in dc.map[i + k:N]:
@@ -15,5 +15,5 @@ def triu(A: dc.float32[N, N]):
 
 
 @dc.program
-def kernel(A: dc.float32[N, N]):
+def kernel(A: dc.float64[N, N]):
     A[:] = np.linalg.cholesky(A) + triu(A, k=1)
