+++ npbench/benchmarks/polybench/jacobi_1d/jacobi_1d_dace.py
@@ -5,7 +5,7 @@ N = dc.symbol('N', dtype=dc.int64)
 
 
 @dc.program
-def kernel(TSTEPS: dc.int64, A: dc.float32[N], B: dc.float32[N]):
+def kernel(TSTEPS: dc.int64, A: dc.float64[N], B: dc.float64[N]):
 
     for t in range(1, TSTEPS):
         B[1:-1] = 0.33333 * (A[:-2] + A[1:-1] + A[2:])
